requirements.txt
functions-framework
google-cloud-bigquery